src/tests/math_net/requirements.txt
langchain-openai
langgraph
langgraph-swarm
python-dotenv
pyyaml
